backend/pyproject.toml
[tool.poetry]
name = "ugaiv2-backend"
version = "0.1.0"
description = "Backend for UGAI analytics platform"
authors = ["Kirill <admin@uga.ai>"]
license = "MIT"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
pydantic = "^2.6.4"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"
black = "^24.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"